Remove polling of data ready flag
The data ready flag should either be polled constantly or the value just
read every measurement interval.
Reason is, that when a new measurement is available, the flag is only
set for 100ms. Now if our clock is imprecise this may lead to a slow
shift which at some point may exceed 100ms, leading to a missed
measurement and a one second delay during polling.